pyproject.toml
[tool.poetry]
name = "rag-poc"
version = "0.1.0"
description = ""
authors = ["marintosti12 <marintosti12@gmail.com>"]
readme = "README.md"
packages = [{include = "src"}] 

[tool.poetry.dependencies]
python = ">=3.12,<3.15"
faiss-cpu = "^1.12.0"
langchain = "^1.0.3"
langchain-community = "^0.4.1"
sentence-transformers = "^5.1.2"
mistralai = "^1.9.11"
python-dotenv = "^1.2.1"
requests = "^2.32.5"
pandas = "^2.3.3"
langchain-mistralai = "^1.0.1"
langchain-core = "^1.0.2"
rich = "^14.2.0"
prompt-toolkit = "^3.0.52"
langchain-huggingface = "^1.0.0"
fastapi = "^0.120.4"
uvicorn = "^0.38.0"
ragas = "^0.3.8"
datasets = "^4.3.0"
langchain-text-splitters = "^1.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.2"
black = "^25.9.0"
pytest-cov = "^7.0.0"
pytest-mock = "^3.15.1"
responses = "^0.25.8"
freezegun = "^1.5.5"
faker = "^38.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
test = "pytest:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=src --cov-report=html --cov-report=term-missing"
markers = [
    "unit: Tests unitaires",
    "integration: Tests d'intégration",
    "slow: Tests lents",
    "api: Tests API"
]